backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy>=2.0.36,<3.0
pydantic>=2.10,<3.0
google-genai
python-dotenv==1.0.1
openpyxl==3.1.2
python-multipart==0.0.9
